move collections client in separate mod
